execinfrapb: move MockDistSQLServer to flowinfra
pb packages are intended to be leaf packages with few or no dependencies.
The mock server in the test utils was pulling in dependencies on packages
that encode business logic, including logic that wants to depend on pb
packages such as pkg/rpc's auth logic that depends on request types.

Release note: none.
Epic: none.
