+++ pkg/sql/colflow/colrpc/BUILD.bazel
@@ -61,6 +61,7 @@ go_test(
         "//pkg/sql/colmem",
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfrapb",
+        "//pkg/sql/flowinfra",
         "//pkg/sql/types",
         "//pkg/testutils",
         "//pkg/util/cancelchecker",
