+++ pkg/sql/execinfrapb/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "expr.go",
         "flow_diagram.go",
         "processors.go",
-        "testutils.go",
     ],
     embed = [":execinfrapb_go_proto"],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/execinfrapb",
@@ -21,9 +20,7 @@ go_library(
     deps = [
         "//pkg/base",
         "//pkg/roachpb",
-        "//pkg/rpc",
         "//pkg/security/username",
-        "//pkg/settings/cluster",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/catalogkeys",
         "//pkg/sql/catalog/catenumpb",
@@ -45,13 +42,9 @@ go_library(
         "//pkg/util/buildutil",
         "//pkg/util/duration",
         "//pkg/util/encoding",
-        "//pkg/util/hlc",
         "//pkg/util/humanizeutil",
-        "//pkg/util/netutil",
         "//pkg/util/optional",
         "//pkg/util/protoutil",
-        "//pkg/util/stop",
-        "//pkg/util/syncutil",
         "//pkg/util/tracing/tracingpb",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
@@ -59,7 +52,6 @@ go_library(
         "@com_github_cockroachdb_redact//:redact",
         "@com_github_dustin_go_humanize//:go-humanize",
         "@com_github_gogo_protobuf//types",
-        "@org_golang_google_grpc//:go_default_library",
     ],
 )
