+++ pkg/sql/flowinfra/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
         "stream_decoder.go",
         "stream_encoder.go",
         "testing_knobs.go",
+        "testutils.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/flowinfra",
     visibility = ["//visibility:public"],
@@ -19,7 +20,9 @@ go_library(
         "//pkg/base",
         "//pkg/kv",
         "//pkg/roachpb",
+        "//pkg/rpc",
         "//pkg/settings",
+        "//pkg/settings/cluster",
         "//pkg/sql/catalog/catenumpb",
         "//pkg/sql/catalog/colinfo",
         "//pkg/sql/execinfra",
@@ -32,23 +35,29 @@ go_library(
         "//pkg/sql/rowenc",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
+        "//pkg/util",
         "//pkg/util/admission",
         "//pkg/util/admission/admissionpb",
         "//pkg/util/buildutil",
         "//pkg/util/cancelchecker",
         "//pkg/util/ctxlog",
+        "//pkg/util/hlc",
         "//pkg/util/log",
         "//pkg/util/mon",
+        "//pkg/util/netutil",
         "//pkg/util/optional",
         "//pkg/util/stop",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
         "//pkg/util/tracing/tracingpb",
+        "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
+        "@com_github_cockroachdb_logtags//:logtags",
         "@com_github_cockroachdb_redact//:redact",
         "@com_github_gogo_protobuf//proto",
         "@io_opentelemetry_go_otel//attribute",
+        "@org_golang_google_grpc//:go_default_library",
     ],
 )
